# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/ai_arts/api/files/download/model/:id": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "download model by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "model id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success  download the code path dir",
                        "schema": {
                            "$ref": "#/definitions/routers.APISuccessResp"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/routers.APIException"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/routers.APIException"
                        }
                    }
                }
            }
        },
        "/ai_arts/api/files/upload/model": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "upload model file, not implemented yet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "upload file key 'data'",
                        "name": "data",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "upload file directory 'dir'",
                        "name": "dir",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/routers.APISuccessResp"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/routers.APIException"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/routers.APIException"
                        }
                    }
                }
            }
        },
...
        "/ai_arts/api/models": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "get model by id",
                "parameters": [
                    {
                        "type": "boolean",
                        "name": "isAdvance",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "orderBy",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "all",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/routers.getModelsetResp"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/routers.APIException"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/routers.APIException"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "create model",
                "parameters": [
                    {
                        "description": "json body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routers.createModelsetReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/routers.APISuccessResp"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/routers.APIException"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/routers.APIException"
                        }
                    }
                }
            }
        },
        "/ai_arts/api/models/:id": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "get model by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "model id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/routers.getModelsetResp"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/routers.APIException"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/routers.APIException"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "update model",
                "parameters": [
                    {
                        "type": "string",
                        "description": "model description",
                        "name": "description",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/routers.APISuccessResp"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/routers.APIException"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/routers.APIException"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "delete model by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "model id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/routers.APISuccessResp"
                        }
                    },
                    "400": {
                        "description": "error",
                        "schema": {
                            "$ref": "#/definitions/routers.APIException"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/routers.APIException"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Modelset": {
            "type": "object",
            "properties": {
                "codePath": {
                    "description": "模型路径",
                    "type": "string"
                },
                "createdAt": {
                    "type": "object",
                    "$ref": "#/definitions/models.UnixTime"
                },
                "creator": {
                    "type": "string"
                },
                "dataFormat": {
                    "type": "string"
                },
                "dataset": {
                    "type": "string"
                },
                "datasetName": {
                    "type": "string"
                },
                "datasetPath": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "object",
                    "$ref": "#/definitions/models.UnixTime"
                },
                "description": {
                    "type": "string"
                },
                "deviceNum": {
                    "type": "integer"
                },
                "deviceType": {
                    "description": "评估设备类型",
                    "type": "string"
                },
                "engine": {
                    "type": "string"
                },
                "evaluationId": {
                    "description": "评估训练任务id",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isAdvance": {
                    "type": "boolean"
                },
                "jobId": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "outputPath": {
                    "description": "输出文件路径",
                    "type": "string"
                },
                "paramPath": {
                    "description": "指定的模型参数路径",
                    "type": "string"
                },
                "params": {
                    "description": "omitempty 值为空，不编码",
                    "type": "object",
                    "$ref": "#/definitions/models.ParamsItem"
                },
                "precision": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "startupFile": {
                    "description": "启动文件路径",
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "object",
                    "$ref": "#/definitions/models.UnixTime"
                },
                "use": {
                    "description": "模型类型 图像分类",
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "models.ParamsItem": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
...
        "models.UnixTime": {
            "type": "object"
        },
...
        "routers.APIException": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "routers.APISuccessResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "routers.createModelsetReq": {
            "type": "object",
            "required": [
                "name",
                "paramPath"
            ],
            "properties": {
                "codePath": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "jobId": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "paramPath": {
                    "type": "string"
                }
            }
        },
...
        "routers.getModelsetResp": {
            "type": "object",
            "properties": {
                "model": {
                    "type": "object",
                    "$ref": "#/definitions/models.Modelset"
                }
            }
        },
...
    }
}